Added protections to classes so they throw errors if needed.

+++ src/main/java/se/iths/vimton/impl/CourseDaoImpl.java
@@ -3,6 +3,7 @@ package se.iths.vimton.impl;
 import se.iths.vimton.dao.CourseDao;
 import se.iths.vimton.entities.Course;
 import se.iths.vimton.entities.Student;
+import se.iths.vimton.entities.Teacher;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -17,14 +18,21 @@ public class CourseDaoImpl implements CourseDao {
 
     @Override
     public void create(Course course) {
-        //todo: !ifExists(course) return;
+        if (exists(course))
+            throw new IllegalArgumentException("Course: " + course.getName() + " already exists.");
         em.getTransaction().begin();
         em.persist(course);
         em.getTransaction().commit();
     }
 
+    private boolean exists(Course course) {
+        return getByName(course.getName()).contains(course);
+    }
+
     @Override
     public void update(Course course) {
+        if(!exists(course))
+            throw new IllegalArgumentException("Could not update course. Course: " + course.getName() + " does not exist.");;
         em.getTransaction().begin();
         em.merge(course);
         em.getTransaction().commit();
@@ -32,7 +40,8 @@ public class CourseDaoImpl implements CourseDao {
 
     @Override
     public void delete(Course course) {
-        //todo: !ifExists(course) throw error
+        if(!exists(course))
+            return;
         em.getTransaction().begin();
         em.remove(course);
         em.getTransaction().commit();
